Fix MonkeyRunner's handling of accessibility ids.
Create a custom class representation for accessibility ids in
IChimpView. This class contains the window id as an int and the
accessibility node id as a long. ChimpView encapsulates the data
from Monkey in this class and passes it on to MonkeyView, which
exposes them to Python.

Change MonkeyDevice to take the ids as strings when
retrieving views by accessibility ids. Jython's ArgParser doesn't
have a parser for long parameters, so pass the value in as a
string and parse the long value in Java.

A change in AccessibilityNodeInfo changed the accessibility
id from int to long and caused errors with MonkeyView and accessibility
ids in MonkeyRunner.

"Adding APIs to enable reporting virtual view hierarchies
to accessibility serivces."
platform/frameworks/base
021078554b902179442a345a9d080a165c3b5139
Svetoslav Ganov <svetoslavganov@google.com>

This commit is related to the following patch:

"Fix AccessibilityNodeInfo view id member reference"
platform/development
6ec4f5d21685c50135337ffad4da2534d6ce423d
Kyle Russell <bkylerussell@gmail.com>

+++ chimpchat/src/main/java/com/android/chimpchat/core/By.java
@@ -36,7 +36,7 @@ public class By {
      * @param accessibilityId the accessibility id of the object you want to select
      * @return a selector that will select the appropriate object by its accessibility ids.
      */
-    public static ISelector accessibilityIds(int windowId, int accessibilityId){
+    public static ISelector accessibilityIds(int windowId, long accessibilityId){
         return new SelectorAccessibilityIds(windowId, accessibilityId);
     }
